Changes so FTL builds with Visual Studio Community 2017 (Version 15.9.4)

+++ tests/functional_tests.cpp
@@ -56,7 +56,7 @@ test_set functional_tests{
 
 				return fequal((f % unary)(2), 4.f/3.f)
 					&& fequal((f % binary)(2,2), 4.f/3.f)
-					&& fequal((unary % binary)(2,2), 8);
+					&& fequal((float)(unary % binary)(2,2), 8.f);
 				;
 			})
 		),
